Refactor annotation package names in preparation for releasing the prototype
