+++ prototype/core/annotation/src/main/java/org/eclipse/sensinact/core/annotation/dto/Data.java
@@ -8,9 +8,9 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
-package org.eclipse.sensinact.prototype.annotation.dto;
+package org.eclipse.sensinact.core.annotation.dto;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -20,14 +20,14 @@ import java.lang.annotation.Target;
 /**
  * Defines a data field in a DTO which maps to a resource. The name of the
  * resource is determined by the following preference order
- * 
+ *
  * <ol>
  * <li>If present, the {@link Resource} annotation present on the field</li>
  * <li>If present, value of the dto field annotated with {@link Resource}</li>
  * <li>If present, the {@link Resource} annotation present on the dto type</li>
  * <li>The name of the field</li>
  * </ol>
- * 
+ *
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.FIELD })
@@ -35,14 +35,14 @@ public @interface Data {
     /**
      * The type of the resource data. If not set then the type of the DTO field is
      * used.
-     * 
+     *
      * @return
      */
     Class<?> type() default Object.class;
 
     /**
      * The resource action when the data field is null
-     * 
+     *
      * @return
      */
     NullAction onNull() default NullAction.IGNORE;
