+++ prototype/core/annotation/src/main/java/org/eclipse/sensinact/core/annotation/dto/Metadata.java
@@ -8,9 +8,9 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
-package org.eclipse.sensinact.prototype.annotation.dto;
+package org.eclipse.sensinact.core.annotation.dto;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -20,7 +20,7 @@ import java.lang.annotation.Target;
 /**
  * Defines a metadata field in a DTO which is applied to a resource. The name of
  * the resource is determined by the following preference order
- * 
+ *
  * <ol>
  * <li>If present, the {@link Resource} annotation present on the field</li>
  * <li>If present, value of the dto field annotated with {@link Resource}</li>
@@ -35,14 +35,14 @@ import java.lang.annotation.Target;
 public @interface Metadata {
     /**
      * The name of the metadata field, if not set then the dto field name is used
-     * 
+     *
      * @return
      */
     String value() default AnnotationConstants.NOT_SET;
 
     /**
      * The resource action when the data field is null
-     * 
+     *
      * @return
      */
     NullAction onNull() default NullAction.IGNORE;
@@ -50,7 +50,7 @@ public @interface Metadata {
     /**
      * The action to take if the annotated field is a map. The default is to treat
      * the map as the value
-     * 
+     *
      * @return
      */
     MapAction[] onMap() default {};
