+++ prototype/core/annotation/src/main/java/org/eclipse/sensinact/core/annotation/dto/Provider.java
@@ -8,9 +8,9 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
-package org.eclipse.sensinact.prototype.annotation.dto;
+package org.eclipse.sensinact.core.annotation.dto;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -18,23 +18,23 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Defines the identity of the model for a given data value
- * 
+ * Defines the name of the provider for a given data value
+ *
  * Either used:
- * 
- * On a String field with no value to supply the model name.
- * 
+ *
+ * On a String field with no value to supply the provider name.
+ *
  * <pre>
- * &#64;Model
- * public String model;
+ * &#64;Provider
+ * public String provider;
  * </pre>
- * 
+ *
  * or
- * 
- * On the type
- * 
+ *
+ * On the type, or a {@link Data} field with a value containing the provider
+ * name
+ *
  * <pre>
- * &#64;Model(&quot;exampleModel&quot;)
  * &#64;Provider(&quot;exampleProvider&quot;)
  * public class MyDto {
  *     &#64;Service(&quot;exampleService&quot;)
@@ -42,14 +42,14 @@ import java.lang.annotation.Target;
  *     public String value;
  * }
  * </pre>
- * 
+ *
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE, ElementType.FIELD })
-public @interface Model {
+public @interface Provider {
     /**
      * The name of the provider
-     * 
+     *
      * @return
      */
     String value() default AnnotationConstants.NOT_SET;
