+++ prototype/core/annotation/src/main/java/org/eclipse/sensinact/core/annotation/dto/Resource.java
@@ -8,9 +8,9 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
-package org.eclipse.sensinact.prototype.annotation.dto;
+package org.eclipse.sensinact.core.annotation.dto;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -18,38 +18,35 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Defines the name of the service for a given data value
- * 
+ * Defines the name of the resource for a given data value
+ *
  * Either used:
- * 
- * On a String field with no value to supply the service name.
- * 
+ *
+ * On a String field with no value to supply the resource name.
+ *
  * <pre>
- * &#64;Service
- * public String service;
+ * &#64;Resource
+ * public String resource;
  * </pre>
- * 
+ *
  * or
- * 
- * On the type, or a {@link Data} field with a value containing the service name
- * 
+ *
+ * On the type, or a {@link Data} field with a value containing the resource
+ * name
+ *
  * <pre>
  * &#64;Provider(&quot;exampleProvider&quot;)
+ * &#64;Service(&quot;exampleService&quot;)
  * public class MyDto {
- *     &#64;Service(&quot;exampleService&quot;)
+ *     &#64;Resource(&quot;exampleResource&quot;)
  *     &#64;Data
  *     public String value;
  * }
  * </pre>
- * 
+ *
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE, ElementType.FIELD })
-public @interface Service {
-    /**
-     * The name of the service
-     * 
-     * @return
-     */
+public @interface Resource {
     String value() default AnnotationConstants.NOT_SET;
 }
