+++ prototype/core/annotation/src/main/java/org/eclipse/sensinact/core/annotation/dto/package-info.java
@@ -11,4 +11,4 @@
 **********************************************************************/
 @org.osgi.annotation.bundle.Export
 @org.osgi.annotation.versioning.Version("0.2.0")
-package org.eclipse.sensinact.prototype.annotation.dto;
+package org.eclipse.sensinact.core.annotation.dto;
