+++ prototype/core/impl/integration-test.bndrun
@@ -57,9 +57,6 @@
 	slf4j.simple;version='[1.7.36,1.7.37)'
 	org.gecko.emf.osgi.api;version='[4.6.0,4.6.1)',\
 	org.gecko.emf.osgi.component;version='[4.6.0,4.6.1)',\
-=======
-	org.gecko.emf.osgi.component;version='[5.0.0,5.0.1)',\
->>>>>>> branch 'model_push' of git@github.com:eclipse/org.eclipse.sensinact.gateway.git
 	org.mockito.junit-jupiter;version='[4.7.0,4.7.1)',\
 	org.mockito.mockito-core;version='[4.7.0,4.7.1)',\
 	org.objenesis;version='[3.2.0,3.2.1)',\
