+++ prototype/examples/pull-based/src/main/java/org/eclipse/sensinact/prototype/pull/_01_SimplePullBasedResource.java
@@ -8,13 +8,13 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
 package org.eclipse.sensinact.prototype.pull;
 
-import org.eclipse.sensinact.prototype.annotation.propertytype.ProviderName;
-import org.eclipse.sensinact.prototype.annotation.propertytype.WhiteboardResource;
-import org.eclipse.sensinact.prototype.annotation.verb.GET;
+import org.eclipse.sensinact.core.annotation.propertytype.ProviderName;
+import org.eclipse.sensinact.core.annotation.propertytype.WhiteboardResource;
+import org.eclipse.sensinact.core.annotation.verb.GET;
 import org.osgi.service.component.annotations.Component;
 
 /**
@@ -27,7 +27,7 @@ public class _01_SimplePullBasedResource {
 
     /**
      * A GET method for a service and resource
-     * 
+     *
      * @return
      */
     @GET(service = "example", resource = "default")
