+++ prototype/examples/push-based/src/main/java/org/eclipse/sensinact/prototype/push/dto/_02_SimpleMultiDTO.java
@@ -8,20 +8,20 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
 package org.eclipse.sensinact.prototype.push.dto;
 
-import org.eclipse.sensinact.prototype.annotation.dto.Data;
-import org.eclipse.sensinact.prototype.annotation.dto.Provider;
-import org.eclipse.sensinact.prototype.annotation.dto.Resource;
-import org.eclipse.sensinact.prototype.annotation.dto.Service;
+import org.eclipse.sensinact.core.annotation.dto.Data;
+import org.eclipse.sensinact.core.annotation.dto.Provider;
+import org.eclipse.sensinact.core.annotation.dto.Resource;
+import org.eclipse.sensinact.core.annotation.dto.Service;
 
 /**
  * This example is a DTO defining two resources with the uris
  * <code>push_example/simple/count</code> and
  * <code>override_example/simple_override/average</code>
- * 
+ *
  * <ul>
  * <li>The provider and service names are defined as annotations, with fields
  * overriding class level</li>
@@ -30,7 +30,7 @@ import org.eclipse.sensinact.prototype.annotation.dto.Service;
  * <li>The resource type is inferred from the data field type unless an
  * annotation defines it.</li>
  * </ul>
- * 
+ *
  */
 @Provider("push_example")
 @Service("simple")
