+++ prototype/examples/push-based/src/main/java/org/eclipse/sensinact/prototype/push/dto/_03_ParameterizedDTO.java
@@ -8,26 +8,26 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*   Kentyou - initial implementation 
+*   Kentyou - initial implementation
 **********************************************************************/
 package org.eclipse.sensinact.prototype.push.dto;
 
-import org.eclipse.sensinact.prototype.annotation.dto.Data;
-import org.eclipse.sensinact.prototype.annotation.dto.Provider;
-import org.eclipse.sensinact.prototype.annotation.dto.Resource;
-import org.eclipse.sensinact.prototype.annotation.dto.Service;
-import org.eclipse.sensinact.prototype.annotation.dto.Timestamp;
+import org.eclipse.sensinact.core.annotation.dto.Data;
+import org.eclipse.sensinact.core.annotation.dto.Provider;
+import org.eclipse.sensinact.core.annotation.dto.Resource;
+import org.eclipse.sensinact.core.annotation.dto.Service;
+import org.eclipse.sensinact.core.annotation.dto.Timestamp;
 
 /**
  * This example is a DTO defining a resource with the uri
  * <code>foo/bar/foobar</code>
- * 
+ *
  * <ul>
  * <li>The provider, service and resource names are defined as annotated
  * fields</li>
  * <li>The timestamp is set using a field</li>
  * </ul>
- * 
+ *
  */
 public class _03_ParameterizedDTO {
